# internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VPC Proof Agent API",
    "version": "v1",
    "description": "Diagnostic and evidence-gathering API for validating a manually provisioned AWS networking environment."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/api/v1/info": {
      "get": {
        "summary": "Agent build info and instance metadata",
        "operationId": "info",
        "responses": {
          "200": {
            "description": "Agent and instance information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InfoResponse" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/probe": {
      "get": {
        "summary": "Full probe report",
        "operationId": "probe",
        "parameters": [
          {
            "name": "X-Force-Refresh",
            "in": "header",
            "required": false,
            "description": "Bypass the probe cache and force a fresh run when true",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "The complete probe report",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProbeReport" }
              }
            }
          }
        }
      }
    },
    "/api/v1/report": {
      "get": {
        "summary": "Full evidence report",
        "operationId": "report",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "Output format",
            "schema": {
              "type": "string",
              "enum": ["json", "markdown", "text"],
              "default": "json"
            }
          },
          {
            "name": "X-Force-Refresh",
            "in": "header",
            "required": false,
            "description": "Bypass the probe cache and force a fresh run when true",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "The rendered evidence report",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReportData" } },
              "text/markdown": { "schema": { "type": "string" } },
              "text/plain": { "schema": { "type": "string" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication; enforced only when auth.enabled is true"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "request_id": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["error", "request_id", "timestamp"]
      },
...
      "AgentInfo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "version": { "type": "string" },
          "commit": { "type": "string" },
          "build_date": { "type": "string" },
          "go_version": { "type": "string" },
          "platform": { "type": "string" }
        }
      },
      "Instance": {
        "type": "object",
        "properties": {
          "instance_id": { "type": "string" },
          "availability_zone": { "type": "string" },
          "private_ip": { "type": "string" },
          "public_ip": { "type": "string" },
          "mac_address": { "type": "string" },
          "subnet_cidr": { "type": "string" },
          "vpc_cidr": { "type": "string" }
        }
      },
      "InfoResponse": {
        "type": "object",
        "properties": {
          "agent": { "$ref": "#/components/schemas/AgentInfo" },
          "instance": { "$ref": "#/components/schemas/Instance" }
        }
      },
      "Summary": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["pass", "fail", "warn", "skip"] },
          "passed": { "type": "integer" },
          "failed": { "type": "integer" },
          "warned": { "type": "integer" },
          "skipped": { "type": "integer" },
          "total": { "type": "integer" },
          "duration": { "type": "integer", "description": "nanoseconds" }
        }
      },
      "Network": {
        "type": "object",
        "properties": {
          "default_gateway": { "type": "string" },
          "default_interface": { "type": "string" },
          "primary_ip": { "type": "string" },
          "dns_addresses": { "type": "string" }
        }
      },
      "ProbeResult": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "status": { "type": "string", "enum": ["pass", "fail", "warn", "skip"] },
          "duration": { "type": "integer", "description": "nanoseconds" },
          "message": { "type": "string" },
          "details": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "hint": { "type": "string" }
        }
      },
      "ProbeReport": {
        "type": "object",
        "properties": {
          "started_at": { "type": "string", "format": "date-time" },
          "duration": { "type": "integer", "description": "nanoseconds" },
          "status": { "type": "string", "enum": ["pass", "fail", "warn", "skip"] },
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProbeResult" }
          }
        }
      },
      "Hint": {
        "type": "object",
        "properties": {
          "rule_id": { "type": "string" },
          "severity": { "type": "string", "enum": ["info", "warning", "critical"] },
          "message": { "type": "string" }
        }
      },
      "ReportData": {
        "type": "object",
        "properties": {
          "agent": { "$ref": "#/components/schemas/AgentInfo" },
          "instance": { "$ref": "#/components/schemas/Instance" },
          "network": { "$ref": "#/components/schemas/Network" },
          "summary": { "$ref": "#/components/schemas/Summary" },
          "probes": { "$ref": "#/components/schemas/ProbeReport" },
          "diagnostics": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Hint" }
          },
          "generated_at": { "type": "string", "format": "date-time" },
          "integrity_hash": { "type": "string", "description": "SHA-256 of the canonical JSON with this field cleared" }
        }
      },
...
    }
  }
}